+++ src/App.js
@@ -5,11 +5,13 @@ import React from 'react'
 import Header from 'modules/Header/Header'
 
 // CSS Imports
+import GlobalStyle from 'src/globalConfig/GlobalStyles'
 import StyledApp from './StyledApp'
 
 const App = () => {
   return (
     <StyledApp>
+      <GlobalStyle />
       <Header />
       <h1>Welcome to React App thats build using Webpack and Babel separately</h1>
     </StyledApp>